# ipx/docs/swagger.yaml
basePath: /
definitions:
...
  v4.BuildNativeEIP1559TransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        type: string
    type: object
  v4.BuildNativeEIP1559TransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
  v4.BuildNativeLegacyTransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        type: string
    type: object
  v4.BuildNativeLegacyTransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
host: localhost:33152
info:
  contact: {}
  description: Ethereum transaction API
  title: evmlab API
  version: "1.0"
paths:
...
  /evm/v4/transaction/native/eip1559:
    post:
      consumes:
      - application/json
      description: Fetches chain state, builds and signs an EIP-1559 native transfer,
        broadcasts it, and returns unsigned_rlp, signed_rlp, tx_hash, r, s, v
      parameters:
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v4.BuildNativeEIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.BuildNativeEIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build, sign, and send EIP-1559 transfer tx
      tags:
      - transaction
  /evm/v4/transaction/native/legacy:
    post:
      consumes:
      - application/json
      description: Fetches chain state, builds and signs an EIP-155 legacy native
        transfer, broadcasts it, and returns unsigned_rlp, signed_rlp, tx_hash, r,
        s, v
      parameters:
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v4.BuildNativeLegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.BuildNativeLegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build, sign, and send legacy transfer tx
      tags:
      - transaction
swagger: "2.0"